+++ src/extension.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
-import { generateTypesCode } from "./code-generator";
-import { askFolder, getCSSModuleFiles, isCSSModule } from "./utils";
+import { addToIgnore, generateTypesCode } from "./code-generator";
+import { askFile, askFolder, getCSSModuleFiles, isCSSModule } from "./utils";
 export function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(
     vscode.workspace.onDidSaveTextDocument(async (e) => {
@@ -18,19 +18,43 @@ export function activate(context: vscode.ExtensionContext) {
         if (uri instanceof vscode.Uri && (await isCSSModule(uri))) {
           await generateTypesCode(uri);
         } else {
-          vscode.window.showErrorMessage('Given URI is not a css module file (*.module.css).');
+          vscode.window.showErrorMessage(
+            "Given URI is not a css module file (*.module.css)."
+          );
         }
       }
     )
   );
   context.subscriptions.push(
     vscode.commands.registerCommand(
-      "css-module-types-and-map.generate-all",
+      "css-module-types-and-map.generate-for-folder",
       async (uri?: vscode.Uri) => {
-        uri = uri instanceof vscode.Uri ? uri : await askFolder(context);
+        uri =
+          uri instanceof vscode.Uri
+            ? uri
+            : await askFolder(context, {
+                openLabel: "Pick a folder for generating code.",
+              });
         if (uri) {
           const files = await getCSSModuleFiles(uri);
           await Promise.all(files.map((file) => generateTypesCode(file)));
+        }
+      }
+    )
+  );
+  context.subscriptions.push(
+    vscode.commands.registerCommand(
+      "css-module-types-and-map.add-to-gitignore",
+      async (uri?: vscode.Uri) => {
+        uri =
+          uri instanceof vscode.Uri
+            ? uri
+            : await askFile({
+                openLabel:
+                  "Pick a '.gitignore' file for adding ignore pattern.",
+              });
+        if (uri) {
+          await addToIgnore(uri, context);
         }
       }
     )